GPT API scripts
ChatGPTHandler.cs file is my attempt at getting the script to work with unity (not working), but in the test folder with the program.cs file I was able to get the API calls working from the command line (need .NET SDK installed to run). Had trouble with figuring out how to get it to work with unity input field and text objects to display the input/output though. Issue was that dragging the object into the input part of a game object running the gpt script wouldnt let the input object be dropped in.